requirements.txt
# Essential Web Framework
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0

# Core NLP Processing
textblob==0.17.1
langdetect==1.0.9

# Google AI Services
google-generativeai>=0.3.1
deep-translator==1.11.4

# Basic utilities (pandas optional - only for advanced file processing)
requests>=2.31.0

# Production Server
gunicorn==21.2.0